+++ Additionals/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Version 1.11.0
 
+* Refactor server/socket config, unify ports and delays ([#15](https://github.com/budul100/Score2Stream/issues/15))
 * Refactor input device matching by name ([#30](https://github.com/budul100/Score2Stream/issues/30))
 
 ## Version 1.10.0